chore: 각 페이지 코드 파일로 분리
