+++ lib/start_page.dart
@@ -1,0 +1,45 @@
+import 'package:flame/components.dart';
+import 'package:flutter/material.dart';
+import 'package:kojaem_novel_game_prototype/route_demo.dart';
+
+class StartPage extends Component with HasGameReference<RouterGame> {
+  StartPage() {
+    addAll([
+      _logo = TextComponent(
+        text: 'Lobby',
+        textRenderer: TextPaint(
+          style: const TextStyle(
+            fontSize: 64,
+            color: Color(0xFFC8FFF5),
+            fontWeight: FontWeight.w800,
+          ),
+        ),
+        anchor: Anchor.center,
+      ),
+      _button1 = RoundedButton(
+        text: 'Level 1',
+        action: () => game.router.pushNamed('level1'),
+        color: const Color(0xffadde6c),
+        borderColor: const Color(0xffedffab),
+      ),
+      _button2 = RoundedButton(
+        text: 'Level 2',
+        action: () => game.router.pushNamed('level2'),
+        color: const Color(0xffdebe6c),
+        borderColor: const Color(0xfffff4c7),
+      ),
+    ]);
+  }
+
+  late final TextComponent _logo;
+  late final RoundedButton _button1;
+  late final RoundedButton _button2;
+
+  @override
+  void onGameResize(Vector2 size) {
+    super.onGameResize(size);
+    _logo.position = Vector2(size.x / 2, size.y / 3);
+    _button1.position = Vector2(size.x / 2, _logo.y + 80);
+    _button2.position = Vector2(size.x / 2, _logo.y + 140);
+  }
+}
